- prevent invalidation when in WM_PAINT - fix a few compiler warnings

+++ vstgui/lib/controls/cautoanimation.cpp
@@ -88,7 +88,7 @@ CAutoAnimation::CAutoAnimation (const CRect& size, CControlListener* listener, i
 	setHeightOfOneImage (heightOfOneImage);
 	totalHeightOfBitmap = heightOfOneImage * getNumSubPixmaps ();
 	setMin (0.f);
-	setMax (totalHeightOfBitmap - (heightOfOneImage + 1));
+	setMax ((float)(totalHeightOfBitmap - (heightOfOneImage + 1.)));
 }
 
 //------------------------------------------------------------------------
